// packages/web/src/api/model/ticketTypeSummary.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 1.0.0
 */

export interface TicketTypeSummary {
  user: number;
  comment: number;
  forum: number;
  blip: number;
  wiki: number;
  pool: number;
  set: number;
  post: number;
  dmail: number;
  replacement: number;
}